Refactor util.
